Change to feed s3 endpoint. (#5)
* Adding overlay

* Changes for usability and readability

Changes to readme per Cheng's suggestion.
Use remainder method on large mouse movements to generate more intuitive behavior (step size 5, mouse move is 13, two movements and remainder is 3 which is saved).
Added placeholder code (displays a checkered background currently) if no data in api endpoint)
Added a hack that reduces the flicker in firefox - puts the previous image to the background so that the delay doesn't make it flicker. (Firefox doesn't fill an image after downloading unless it is displaying).

* Formatting

* Changing to pull lists of spinner images from feed server rather than api server.
Also removes all dependency on JQuery

File end of line.